# backend/openapi.yaml
openapi: 3.0.3
info:
  title: MediTime API
  description: |
    API REST complète pour **MediTime**, application de gestion de médicaments et suivi des traitements médicaux.

    ## 🔐 Authentification
    L'API utilise **Supabase Auth** avec tokens JWT Bearer.  
    La plupart des endpoints nécessitent le header `Authorization: Bearer <token>`.

    ### Obtenir un token
    1. Authentifiez-vous via Supabase Auth (email/password, Google, etc.)
    2. Récupérez le token JWT depuis la session Supabase
    3. Ajoutez-le au header : `Authorization: Bearer <votre_token>`

    ## 🏗 Architecture
    | Composant | Technologie |
    |-----------|-------------|
    | Frontend | React 19 (Vite) + Bootstrap 5 + Supabase Client |
    | Backend | Flask (Python) + PostgreSQL + Raw SQL (psycopg2) |
    | Base de données | PostgreSQL avec Row Level Security (RLS) |
    | Infrastructure | Docker + CapRover + Render |

    ## 🚀 Guide de démarrage
    1. `GET /api/status` — Vérifier la connexion
    2. `GET /api/user/sync` — Synchroniser votre profil
    3. `GET /api/calendars` — Lister vos calendriers
    4. `POST /api/calendars` — Créer un calendrier

    ## 📦 Codes de réponse
    | Code | Signification |
    |------|---------------|
    | 200 | Succès |
    | 400 | Requête invalide (données manquantes) |
    | 401 | Non authentifié |
    | 403 | Accès refusé |
    | 404 | Ressource non trouvée |
    | 500 | Erreur serveur |

  version: 2.0.0
  contact:
    name: MediTime Support
...
  license:
...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT Supabase Auth.
        
        **Format du header :**
        ```
        Authorization: Bearer eyJhbGciOiJIUzI1NiIs...
        ```
  schemas:
    Error:
      type: object
      required: [message, code]
      properties:
        message:
          type: string
          description: Message d'erreur lisible
          example: "Erreur lors de la récupération des données"
        code:
          type: string
          description: Code d'erreur machine
          example: "CALENDAR_FETCH_ERROR"
        i18n_key:
          type: string
          description: Clé de traduction i18n
          example: "api.calendar.fetch_error"
        error:
          type: string
          description: Détails techniques (dev uniquement)
        status_code:
          type: integer
          example: 500
    Success:
      type: object
      required: [message, code]
      properties:
        message:
          type: string
          example: "Opération réussie"
        code:
          type: string
          example: "OPERATION_SUCCESS"
        i18n_key:
          type: string
        data:
          type: object
          additionalProperties: true
...
    Calendar:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          example: "Mon traitement"
        owner_uid:
          type: string
          format: uuid
        boxes_count:
          type: integer
          description: Nombre de boîtes
        ifLowStock:
          type: boolean
          description: Au moins une boîte en stock faible
        stock_decrement_method:
          type: string
          enum: [auto, manual]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
...
    SharedCalendar:
      type: object
      properties:
        id:
          type: integer
        access:
          type: string
          enum: [view, edit]
        name:
          type: string
        owner_uid:
          type: string
          format: uuid
        owner_name:
          type: string
        owner_email:
          type: string
        owner_photo_url:
          type: string
        notifications_enabled:
          type: boolean
        boxes_count:
          type: integer
        ifLowStock:
          type: boolean
...
    Schedule:
      type: object
      properties:
        schedule:
          type: object
          additionalProperties:
            type: array
            items:
              type: object
        table:
          type: object
        calendar_name:
          type: string
        if_low_stock:
          type: boolean
paths:
...
  /calendars:
    get:
      tags: [Calendars]
      summary: Lister mes calendriers
      operationId: getCalendars
      responses:
        '200':
          description: Calendriers récupérés
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          calendars:
                            type: array
                            items:
                              $ref: '#/components/schemas/Calendar'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Calendars]
      summary: Créer un calendrier
      operationId: createCalendar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [calendarName]
              properties:
                calendarName:
                  type: string
                  example: "Mon traitement"
      responses:
        '200':
          description: Calendrier créé
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          calendarId:
                            type: integer
                          calendarName:
                            type: string
        '400':
          description: Nom manquant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /shared/users/calendars:
    get:
      tags: [Shared Calendars]
      summary: Lister les calendriers partagés avec moi
      operationId: getSharedCalendars
      responses:
        '200':
          description: Calendriers partagés récupérés
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          calendars:
                            type: array
                            items:
                              $ref: '#/components/schemas/SharedCalendar'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /shared/grouped:
    get:
      tags: [Shared Calendars]
      summary: Vue groupée des calendriers partagés
      operationId: getSharedGrouped
      responses:
        '200':
          description: Vue groupée récupérée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /shared/users/calendars/{calendar_id}:
    delete:
      tags: [Shared Calendars]
      summary: Quitter un calendrier partagé
      operationId: leaveSharedCalendar
      parameters:
        - name: calendar_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Calendrier quitté
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '403':
          description: Accès non autorisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /shared/users/calendars/{calendar_id}/schedule:
    get:
      tags: [Shared Calendars]
      summary: Planning d'un calendrier partagé
      operationId: getSharedCalendarSchedule
      parameters:
        - name: calendar_id
          in: path
          required: true
          schema:
            type: integer
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Planning récupéré
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Schedule'
        '403':
          description: Accès non autorisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /shared/users/calendars/{calendar_id}/notifications:
    get:
      tags: [Shared Calendars]
      summary: Statut notifications calendrier partagé
      operationId: getSharedCalendarNotifications
      parameters:
        - name: calendar_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Statut récupéré
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          notifications-enabled:
                            type: boolean
        '404':
          description: Non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags: [Shared Calendars]
      summary: Basculer notifications calendrier partagé
      operationId: toggleSharedCalendarNotifications
      parameters:
        - name: calendar_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Notifications basculées
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /shared/users/calendars/{calendar_id}/stock-decrement-method:
    get:
      tags: [Shared Calendars]
      summary: Méthode décompte stock (partagé)
      operationId: getSharedStockDecrementMethod
      parameters:
        - name: calendar_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Méthode récupérée
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          method:
                            type: string
                            enum: [auto, manual]
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...